tests/wpt/web-platform-tests/css/cssom/css-style-attribute-modifications.html

<!DOCTYPE html>
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:ecobos@igalia.com">
<link rel="help" href="https://drafts.csswg.org/cssom/#the-elementcssinlinestyle-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="test" style="color: red"></div>
<script>
test(function() {
  var el = document.getElementById("test");
  el.style.color = "";
  assert_true(el.hasAttribute("style"));

  el.removeAttribute("style");
  assert_false(el.hasAttribute("style"));
}, "Mutating the style declaration doesn't remove the style attribute");
</script>
